+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Fix: Adding related products to cart.
     * Fix: Fatal error related to wrong namespace use.
     * Fix: Allow to query product tags and categories at the same time.
+    * Fix: Adding custom settings tab by using Integration.
 * 2.1 - 2017.04.26:
     * New: Admin notices.
     * New: My Downloads panel in my account.